requirements.txt
websockets>=15.0
pyyaml>=6.0
jsonschema>=4.20
